+++ R/get_apicarto_codes_postaux.R
@@ -19,7 +19,13 @@
 #'
 #' info_commune <- get_apicarto_codes_postaux("29760")
 #'
-#' code_post <- c("29760", "29260")
+#' code_post <- c("29760", "08170")
+#' info_communes <- get_apicarto_codes_postaux(code_post)
+#'
+#' code_post <- c("12345")
+#' info_communes <- get_apicarto_codes_postaux(code_post)
+#'
+#' code_post <- c("12345", "08170")
 #' info_communes <- get_apicarto_codes_postaux(code_post)
 #'}
 #'
@@ -29,19 +35,30 @@
 
 get_apicarto_codes_postaux <- function(code_post){
 
-   hit_code_post <- function(code_post){
-      resp <- build_req(path = paste0("api/codes-postaux/communes/",
-                                      code_post)) |>
-         req_perform() |>
-         resp_body_json()
-      resp <- do.call(rbind, resp)
-   }
+   pad0 <- \(x, n) if (is.null(x)) NULL else gsub(" ", "0", sprintf(paste0("%", n, "s"), x))
+   code_post <- pad0(code_post, 5) |> unique()
 
-   resp <- Map(hit_code_post,
-               code_post)
+   fetch_data <- function(code_post){
+      tryCatch({
+         req <- request("https://apicarto.ign.fr") |>
+            req_url_path("api/codes-postaux/communes") |>
+            req_url_path_append(code_post) |>
+            req_user_agent("happign (https://paul-carteron.github.io/happign/)") |>
+            req_options(ssl_verifypeer = 0) |>
+            req_perform() |>
+            resp_body_json()
+      },
+      error = function(e) {
+         warning("No data found for : ", code_post, call. = F)
+         return(NULL)
+      })}
 
-   # bind rows of each Map call
-   resp <- data.frame(do.call(rbind, resp))
+   resp <- lapply(code_post, fetch_data)
+   resp <- Filter(Negate(is.null), resp)
 
-   return(resp)
+   df <- do.call(rbind, lapply(resp, function(x) {
+      do.call(rbind.data.frame, x)
+   }))
+
+   return(df)
 }