+++ tests/testthat/apicarto-codes-postaux/apicarto.ign.fr/api/codes-postaux/communes/01234.R
@@ -1,0 +1,8 @@
+structure(list(method = "GET", url = "https://apicarto.ign.fr/api/codes-postaux/communes/01234", 
+    status_code = 404L, headers = structure(list(date = "Thu, 14 Aug 2025 12:21:10 GMT", 
+        `content-type` = "text/plain; charset=utf-8", `content-length` = "9", 
+        `x-powered-by` = "Express", `access-control-allow-origin` = "*", 
+        `cache-control` = "private, no-cache, no-store, must-revalidate", 
+        expires = "-1", pragma = "no-cache", etag = "W/\"9-0gXL1ngzMqISxa6S1zx3F4wtLyg\"", 
+        `strict-transport-security` = "max-age=31536000; includeSubDomains"), class = "httr2_headers"), 
+    body = charToRaw("Not Found"), cache = new.env(parent = emptyenv())), class = "httr2_response")