+++ tests/testthat/get_api_codes_postaux/apicarto.ign.fr/api/codes-postaux/communes/29160.R
@@ -1,8 +1,0 @@
-structure(list(method = "GET", url = "https://apicarto.ign.fr/api/codes-postaux/communes/29160", 
-    status_code = 200L, headers = structure(list(`Access-Control-Allow-Origin` = "*", 
-        `Cache-Control` = "private, no-cache, no-store, must-revalidate", 
-        Expires = "-1", Pragma = "no-cache", `Content-Type` = "application/json; charset=utf-8", 
-        `Content-Length` = "200", Date = "Thu, 03 Aug 2023 10:23:21 GMT", 
-        Connection = "keep-alive", `Keep-Alive` = "timeout=5", 
-        `Set-Cookie` = "REDACTED", `Strict-Transport-Security` = "max-age=31536000; includeSubDomains"), class = "httr2_headers"), 
-    body = charToRaw("[{\"codePostal\":\"29160\",\"codeCommune\":\"29042\",\"nomCommune\":\"Crozon\",\"libelleAcheminement\":\"CROZON\"},{\"codePostal\":\"29160\",\"codeCommune\":\"29120\",\"nomCommune\":\"Lanvéoc\",\"libelleAcheminement\":\"LANVEOC\"}]")), class = "httr2_response")